Refactoring for checks on strings containing formatted numeric values when used in mathematical operations in the Calculation Engine
