+++ src/PhpSpreadsheet/Shared/StringHelper.php
@@ -2,17 +2,8 @@
 
 namespace PhpOffice\PhpSpreadsheet\Shared;
 
-use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
-
 class StringHelper
 {
-    /**    Constants                */
-    /**    Regular Expressions        */
-    //    Fraction
-    const STRING_REGEXP_FRACTION = '~^\s*(-?)((\d*)\s+)?(\d+\/\d+)\s*$~';
-
-    const STRING_REGEXP_PERCENT = '~^(?:(?: *(?<PrefixedSign>[-+])? *\% *(?<PrefixedSign2>[-+])? *(?<PrefixedValue>[0-9]+\.?[0-9*]*(?:E[-+]?[0-9]*)?) *)|(?: *(?<PostfixedSign>[-+])? *(?<PostfixedValue>[0-9]+\.?[0-9]*(?:E[-+]?[0-9]*)?) *\% *))$~i';
-
     /**
      * Control characters array.
      *
@@ -538,46 +529,6 @@ class StringHelper
         }
 
         return implode('', $characters);
-    }
-
-    /**
-     * Identify whether a string contains a fractional numeric value,
-     * and convert it to a numeric if it is.
-     *
-     * @param string $operand string value to test
-     */
-    public static function convertToNumberIfFraction(string &$operand): bool
-    {
-        if (preg_match(self::STRING_REGEXP_FRACTION, $operand, $match)) {
-            $sign = ($match[1] == '-') ? '-' : '+';
-            $wholePart = ($match[3] === '') ? '' : ($sign . $match[3]);
-            $fractionFormula = '=' . $wholePart . $sign . $match[4];
-            $operand = Calculation::getInstance()->_calculateFormulaValue($fractionFormula);
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Identify whether a string contains a percentage, and if so,
-     * convert it to a numeric.
-     *
-     * @param string $operand string value to test
-     */
-    public static function convertToNumberIfPercent(string &$operand): bool
-    {
-        $match = [];
-        if (preg_match(self::STRING_REGEXP_PERCENT, $operand, $match, PREG_UNMATCHED_AS_NULL)) {
-            //Calculate the percentage
-            $sign = ($match['PrefixedSign'] ?? $match['PrefixedSign2'] ?? $match['PostfixedSign']) ?? '';
-            $operand = (float) ($sign . ($match['PostfixedValue'] ?? $match['PrefixedValue'])) / 100;
-
-            return true;
-        }
-
-        return false;
     }
 
     /**
